requirements.txt
aiogram==3.15.0
openai==1.57.2
llama-index==0.12.5
python-dotenv==1.0.1
txt2docx==1.0.2